backend/requirements.txt
fastapi
uvicorn
python-dotenv
huggingface_hub
chromadb
sentence-transformers
python-multipart
pypdf
python-docx
requests
jinja2
httpx
sqlmodel
gunicorn

